requirements.txt
tensorflow==2.15.0
keras==2.15.0
numpy==1.24.3
matplotlib==3.7.1
scikit-learn==1.2.2
Pillow==9.5.0
streamlit==1.22.0
opencv-python==4.7.0.72
pandas==2.0.1
tqdm==4.65.0
kaggle==1.5.13 